(chore) codebase cleanup + readme update
