+++ src/components/mod.rs
@@ -2,11 +2,9 @@ mod dump;
 mod sidebar;
 mod taskbar;
 mod language_icon;
-mod diff_modal;
 
 pub(crate) use sidebar::SideBar;
 pub(crate) use dump::DumpItem;
 pub(crate) use dump::DumpGroup;
 pub(crate) use taskbar::Taskbar;
 pub(crate) use language_icon::LanguageIcon;
-pub(crate) use diff_modal::DiffModal;
