+++ src/modals/mod.rs
@@ -1,0 +1,5 @@
+mod settings_modal;
+mod diff_modal;
+
+pub use settings_modal::SettingsModal;
+pub use diff_modal::DiffModal;
